[HF] Fix warnings (#47)
* few changes made to fix Wconversion warnings

+++ tests/stl/open_map_check.cpp
@@ -287,8 +287,8 @@ TEST(open_map_test, test_rehash) {
 	map[226] = 2216;
 	map[337] = 2317;
 	map[448] = 2418;
-	int keys[] = {0, 1, 2, 3, 4, 115, 226, 337, 448};
-	int values[] = {0, 10, 20, 30, 40, 2115, 2216, 2317, 2418};
+	uint16_t keys[] = {0, 1, 2, 3, 4, 115, 226, 337, 448};
+	uint16_t values[] = {0, 10, 20, 30, 40, 2115, 2216, 2317, 2418};
 	for (uint16_t i = 0; i < 9; i++) {
 		ASSERT_TRUE(map.contains(keys[i]));
 		ASSERT_EQ(*map.find(keys[i]), values[i]);
